OPEN - task 23: Remove old unused parts of API http://github.com/triceo/ras12/issues/issue/23

+++ ras2012-solver/src/main/java/org/drools/planner/examples/ras2012/model/Route.java
@@ -186,7 +186,7 @@ public class Route extends Visualizable implements Comparable<Route>, Directed {
                 final BigDecimal length = a.getLength();
                 final int speed = this.isEastbound() ? a.getTrack().getSpeedEastbound() : a
                         .getTrack().getSpeedWestbound();
-                result += Converter.getTimeFromSpeedAndDistance(speed, length);
+                result += Converter.getTimeFromSpeedAndDistance(BigDecimal.valueOf(speed), length);
             }
             this.travellingTimeInMilliseconds = result;
         }
